Created a dynamic list of reviews /no css/ and started up test for enzyme-jest
