+++ _tests_/app.testing.js
@@ -1,0 +1,2 @@
+import React from 'react';
+import { shallow, mount, render } from 'enzyme';
