+++ package.json
@@ -7,6 +7,10 @@
   "engines": {
     "node": ">=6.13.0"
   },
+  "jest": {
+    "verbose": true,
+    "setupFilesAfterEnv": ["<rootDir>src/setupTests.js"]
+  },
   "scripts": {
     "seed": "node seed.js",
     "react-dev": "webpack -d --watch",
@@ -28,6 +32,7 @@
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.6",
     "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.14.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "webpack": "^4.39.2",
     "webpack-cli": "^3.3.7"
