+++ server/index.js
@@ -1,17 +1,17 @@
 const db = require('../db');
 const express = require('express');
-const bodyParser = require('body-parser');
 
 const app = express();
 const PORT = 3000;
 
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({extended: true}));
-
-app.use(express.static(__dirname + '../client/dist'));
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+app.use(express.static(`${__dirname}./../client/dist`));
 
 app.get('/api/reviews/:id', (req, res) => {
-  db.ReviewComment.find({listingID: req.params.id}, (error, data) => {
+  console.log(req.params.id);
+
+  db.ReviewComment.find({listingID: `${req.params.id}`}, (error, data) => {
     if(error) {
       res.status(500).send(error);
     } else {
